Refactor CommunityInterest to component
